refac + feat: product card component + add or edit product modal + edit product feature
