+++ types/IProduct.ts
@@ -1,0 +1,7 @@
+export default interface IProduct {
+  id: number;
+  name: string;
+  amount: number;
+  unit: "KG" | "L";
+  updatedDate?: string | null;
+}
